+++ CHANGES.md
@@ -24,6 +24,7 @@
 
 - Config file
   - [BREAKING CHANGE - TODO] Change default parent key to `"parent.filenames"`
+  - Fix issue where when reading a config file without a parent leads to argument passing error
 - Config I/O protocol
   - [BREAKING CHANGE] ConfigIO support for writing to `java.io.Writer` (besides `java.io.OutputStream`)
 - Lookup function